// server/vance-addon-brain-kanban/client/src/generated/kanban/KanbanBoardView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.kanban.KanbanBoardView
*/

import { KanbanArtefactSummary } from './KanbanArtefactSummary';
import { KanbanColumnView } from './KanbanColumnView';
import { KanbanCardView } from './KanbanCardView';

export interface KanbanBoardView {
  folder: string;
  manifestPath: string;
  title?: string;
  description?: string;
  wipEnforce: string;
  boardStyle: string;
  columns: KanbanColumnView[];
  cards: KanbanCardView[];
  artefacts: KanbanArtefactSummary[];
}