// server/vance-addon-brain-kanban/client/src/generated/kanban/KanbanCardCreateRequest.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.kanban.KanbanCardCreateRequest
*/

export interface KanbanCardCreateRequest {
  title: string;
  column?: string;
  priority?: string;
  assignee?: string;
  labels: string[];
  dueDate?: string;
  estimate?: number;
  blocked: boolean;
  body?: string;
  filename?: string;
}