// server/vance-addon-brain-kanban/client/src/generated/kanban/KanbanRebuildResponse.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.kanban.KanbanRebuildResponse
*/

import { KanbanArtefactSummary } from './KanbanArtefactSummary';

export interface KanbanRebuildResponse {
  folder: string;
  artefacts: KanbanArtefactSummary[];
}